fix: Propagate all inferred properties to a function argument
Each function arguments gets a fresh type variable assigned to them
to which a type gets inferred to in the function body. This then does not
get propagated to the type (and constraint) created for the function
itself as the substitution(s) are not returned to the
`FunctionExpr::infer` call-site.

This is a hack that gets the test passing and I don't know if this is
actually sound, there is probably a better solution for this out there
with some help.

Closes #4051